+++ library/src/main/java/br/liveo/adapter/NavigationLiveoAdapter.java
@@ -22,6 +22,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import java.util.List;
@@ -31,18 +33,23 @@ import br.liveo.navigationliveo.R;
 
 public class NavigationLiveoAdapter extends BaseAdapter {
 
+    private int mColorIcon = 0;
+    private int mColorName = 0;
+
     private int mNewDrawable = 0;
     private int mColorDefault = 0;
 	private final Context mcontext;
     private boolean mRemoveAlpha = false;
 	private final List<NavigationLiveoItemAdapter> mList;
 	
-	public NavigationLiveoAdapter(Context context, List<NavigationLiveoItemAdapter> list, int drawable, int colorDefault, boolean removeAlpha) {
+	public NavigationLiveoAdapter(Context context, List<NavigationLiveoItemAdapter> list, int drawable, int colorDefault, boolean removeAlpha, int colorIcon, int colorName) {
 		this.mList = list;		
 		this.mcontext = context;
-        this.mColorDefault = colorDefault;
+        this.mColorIcon = colorIcon;
+        this.mColorName = colorName;
         this.mNewDrawable = drawable;
         this.mRemoveAlpha = removeAlpha;
+        this.mColorDefault = colorDefault;
 	}
 
 	@Override
@@ -107,6 +114,9 @@ public class NavigationLiveoAdapter extends BaseAdapter {
 		public ImageView icon;
         public TextView counter;
 
+        public RelativeLayout layoutDados;
+        public LinearLayout layoutSeparator;
+
 		public ViewHolder(){
 		}
 	}
@@ -121,25 +131,23 @@ public class NavigationLiveoAdapter extends BaseAdapter {
 
 	public View getView(int position, View convertView, ViewGroup parent) {
 
+        ViewHolder holder;
 		NavigationLiveoItemAdapter item = mList.get(position);
-		ViewHolder holder;
-		
+
 		if (convertView == null) {
 			holder = new ViewHolder();
 			
-			int layout = ((item.isHeader) ? (item.title != null && !item.title.trim().equals(""))
-                   ?                                               R.layout.navigation_list_item_sub_header
-                                                                  :R.layout.navigation_list_item_sub_header_line
+			int layout = R.layout.navigation_list_item_icon;
 
+            LayoutInflater inflater = (LayoutInflater) mcontext.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+            convertView = inflater.inflate(layout, parent, false);
 
-					                      : (item.icon != 0) ? R.layout.navigation_list_item_icon :
-                                                               R.layout.navigation_list_item);
-			
-			convertView = LayoutInflater.from(mcontext).inflate(layout, null);			
-			
 			holder.title = (TextView) convertView.findViewById(R.id.title);
 			holder.counter = (TextView) convertView.findViewById(R.id.counter);
 			holder.icon = (ImageView) convertView.findViewById(R.id.icon);
+
+            holder.layoutDados = (RelativeLayout) convertView.findViewById(R.id.layoutDados);
+            holder.layoutSeparator = (LinearLayout) convertView.findViewById(R.id.layoutSeparator);
 
             convertView.setTag(holder);
 		}else{
@@ -147,14 +155,27 @@ public class NavigationLiveoAdapter extends BaseAdapter {
 		}
 
 		if (holder.title != null){
-			holder.title.setText(item.title);
+
+            holder.title.setText(item.title);
 
             if (!item.isHeader) {
+                holder.layoutSeparator.setVisibility(View.GONE);
+                holder.layoutDados.setVisibility(View.VISIBLE);
                 setAlpha(holder.title, (item.checked ? 1f : 0.87f));
 
                 holder.title.setTextColor((!item.isHeader && item.checked && item.colorSelected > 0 ?
                         mcontext.getResources().getColor(item.colorSelected) :
+                        mColorName > 0 ? mcontext.getResources().getColor(mColorName) :
                         mcontext.getResources().getColor(R.color.nliveo_black)));
+            }else{
+                holder.layoutSeparator.setVisibility(View.VISIBLE);
+
+                if (item.title.equals("")){
+                    holder.title.setVisibility(View.GONE);
+                    holder.layoutDados.setVisibility(View.GONE);
+                }else{
+                    holder.layoutDados.setVisibility(View.VISIBLE);
+                }
             }
 		}
 
@@ -180,6 +201,7 @@ public class NavigationLiveoAdapter extends BaseAdapter {
                 holder.icon.setColorFilter((!item.isHeader && item.checked && item.colorSelected > 0 ?
                         mcontext.getResources().getColor(item.colorSelected) :
                         (mColorDefault != 0 ? mcontext.getResources().getColor(mColorDefault) :
+                                mColorIcon > 0 ? mcontext.getResources().getColor(mColorIcon) :
                                            mcontext.getResources().getColor(R.color.nliveo_black))));
 			} else {
 				holder.icon.setVisibility(View.GONE);